[REFACTOR] Code Convention, Chatroom, Fields 수정 (#176)
* refactor: code convention, chatroom, fields added

* refactor: concludePrice 나중에 추가

* refactor: concludePrice auction에서는 삭제 전체조회할 때는 일단은 넣어둠

* refactor: json format info에서 삭제

* refactor: code review 후 수정된것

* refactor: 필요없는 코드 삭제

* refactor: method명 수정

* refactor: reader 이름 수정

+++ omocha-domain/src/main/java/org/omocha/domain/auction/Auction.java
@@ -48,7 +48,7 @@ public class Auction extends BaseEntity {
 
 	private Long nowPrice;
 
-	private Integer bidCount;
+	private Long bidCount;
 
 	private Long bidUnit;
 
@@ -73,7 +73,7 @@ public class Auction extends BaseEntity {
 		String content,
 		Long startPrice,
 		Long nowPrice,
-		Integer bidCount,
+		Long bidCount,
 		Long bidUnit,
 		String thumbnailPath,
 		LocalDateTime startDate,
